SeriwDetails, SerieList, SerieListByGrid sayfaları oluşturuldu.

+++ EntityLayer/Concrete/Serie.cs
@@ -22,7 +22,7 @@ namespace EntityLayer.Concrete
 
         public string Image { get; set; }
 
-        public int Star { get; set; }
+        public float Star { get; set; }
 
         public DateOnly Date { get; set; }
         public int Season { get; set; }
